+++ src/commands/hardis/org/diagnose/audittrail.ts
@@ -28,6 +28,9 @@ export default class DiagnoseAuditTrail extends SfdxCommand {
 
 Regular setup actions performed in major orgs are filtered.
 
+- ""
+  - createScratchOrg
+  - deleteScratchOrg
 - Certificate and Key Management
   - insertCertificate
 - Groups
@@ -155,6 +158,10 @@ monitoringAllowedSectionsActions:
     }
 
     this.allowedSectionsActions = {
+      "": [
+        "createScratchOrg",
+        "deleteScratchOrg"
+      ],
       "Certificate and Key Management": ["insertCertificate"],
       Groups: ["groupMembership"],
       "Manage Users": [
@@ -214,17 +221,18 @@ monitoringAllowedSectionsActions:
     let suspectUsers = [];
     let suspectActions = [];
     const auditTrailRecords = queryRes.records.map((record) => {
+      const section = record?.Section || "";
       record.Suspect = false;
       // Unallowed actions
       if (
-        (this.allowedSectionsActions[record.Section] && !this.allowedSectionsActions[record.Section].includes(record.Action)) ||
-        !this.allowedSectionsActions[record.Section]
+        (this.allowedSectionsActions[section] && !this.allowedSectionsActions[section].includes(record.Action)) ||
+        !this.allowedSectionsActions[section]
       ) {
         record.Suspect = true;
-        record.SuspectReason = `Manual config in unallowed section ${record.Section} with action ${record.Action}`;
+        record.SuspectReason = `Manual config in unallowed section ${section} with action ${record.Action}`;
         suspectRecords.push(record);
         suspectUsers.push(record["CreatedBy.Username"]);
-        suspectActions.push(`${record.Section} - ${record.Action}`);
+        suspectActions.push(`${section} - ${record.Action}`);
         return record;
       }
       return record;